+++ R/tree.R
@@ -28,6 +28,7 @@ NULL
 #' @param reorder.numeric Re-order identity classes according to position on
 #' the tree, assigning a numeric value ('1' is the leftmost node)
 #' @param verbose Show progress updates
+#' @inheritParams AverageExpression
 #'
 #' @return A Seurat object where the cluster tree can be accessed with \code{\link{Tool}}
 #'
@@ -49,6 +50,7 @@ BuildClusterTree <- function(
   dims = NULL,
   graph = NULL,
   # do.plot = TRUE,
+  slot = 'data',
   reorder = FALSE,
   reorder.numeric = FALSE,
   verbose = TRUE
@@ -117,6 +119,7 @@ BuildClusterTree <- function(
     data.avg <- AverageExpression(
       object = object,
       features = features,
+      slot = slot,
       verbose = verbose
     )[[1]]
     data.dist <- dist(x = t(x = data.avg[features, ]))
@@ -127,13 +130,14 @@ BuildClusterTree <- function(
     if (verbose) {
       message("Reordering identity classes and rebuilding tree")
     }
-    old.ident.order <- sort(x = levels(x = object))
+    old.ident.order <- levels(x = object)
     data.tree <- Tool(object = object, slot = 'BuildClusterTree')
     all.desc <- GetDescendants(tree = data.tree, node = (data.tree$Nnode + 2))
     all.desc <- old.ident.order[all.desc[all.desc <= (data.tree$Nnode + 1)]]
-    levels(x = object) <- all.desc
+    Idents(object = object) <- factor(x = Idents(object = object), levels = all.desc, ordered = TRUE)
     if (reorder.numeric) {
-      Idents(object = object) <- as.integer(x = Idents(object = object))
+      new.levels <- sort(x = unique(x = as.integer(x = Idents(object = object))))
+      Idents(object = object) <- factor(x = as.integer(x = Idents(object = object)), levels = new.levels)
       object[['tree.ident']] <- as.integer(x = Idents(object = object))
     }
     object <- BuildClusterTree(
@@ -141,6 +145,7 @@ BuildClusterTree <- function(
       features = features,
       dims = dims,
       graph = graph,
+      slot = slot,
       reorder = FALSE,
       verbose = verbose
     )