+++ R/visualization.R
@@ -188,6 +188,10 @@ DimHeatmap <- function(
 #' @param raster If true, plot with geom_raster, else use geom_tile. geom_raster may look blurry on
 #' some viewing applications such as Preview due to how the raster is interpolated. Set this to FALSE
 #' if you are encountering that issue (note that plots may take longer to produce/render).
+#' @param draw.lines Include white lines to separate the groups
+#' @param lines.width Integer number to adjust the width of the separating white lines.
+#' Corresponds to the number of "cells" between each group.
+#' @param group.bar.height Scale the height of the color bar
 #' @param combine Combine plots into a single gg object; note that if TRUE; themeing will not work
 #' when plotting multiple dimensions
 #'
@@ -216,6 +220,9 @@ DoHeatmap <- function(
   hjust = 0,
   angle = 45,
   raster = TRUE,
+  draw.lines = TRUE,
+  lines.width = NULL,
+  group.bar.height = 0.02,
   combine = TRUE
 ) {
   cells <- cells %||% colnames(x = object)
@@ -257,11 +264,37 @@ DoHeatmap <- function(
   # group.use <- factor(x = group.use[cells])
   plots <- vector(mode = 'list', length = ncol(x = groups.use))
   for (i in 1:ncol(x = groups.use)) {
+    data.group <- data
     group.use <- groups.use[, i, drop = TRUE]
-    group.use <- factor(x = group.use)
+    if (!is.factor(x = group.use)) {
+      group.use <- factor(x = group.use)
+    }
     names(x = group.use) <- cells
+    if (draw.lines) {
+      # create fake cells to serve as the white lines, fill with NAs
+      lines.width <- lines.width %||% ceiling(x = nrow(x = data.group) * 0.0025)
+      placeholder.cells <- sapply(
+        X = 1:(length(x = levels(x = group.use)) * lines.width),
+        FUN = function(x) {
+          return(RandomName(length = 20))
+        }
+      )
+      placeholder.groups <- rep(x = levels(x = group.use), times = lines.width)
+      group.levels <- levels(x = group.use)
+      names(x = placeholder.groups) <- placeholder.cells
+      group.use <- as.vector(x = group.use)
+      names(x = group.use) <- cells
+      group.use <- factor(x = c(group.use, placeholder.groups), levels = group.levels)
+      na.data.group <- matrix(
+        data = NA,
+        nrow = length(x = placeholder.cells),
+        ncol = ncol(x = data.group),
+        dimnames = list(placeholder.cells, colnames(x = data.group))
+      )
+      data.group <- rbind(data.group, na.data.group)
+    }
     plot <- SingleRasterMap(
-      data = data,
+      data = data.group,
       raster = raster,
       disp.min = disp.min,
       disp.max = disp.max,
@@ -271,13 +304,24 @@ DoHeatmap <- function(
     )
     if (group.bar) {
       # TODO: Change group.bar to annotation.bar
+      group.use2 <- sort(x = group.use)
+      if (draw.lines) {
+        na.group <- RandomName(length = 20)
+        levels(x = group.use2) <- c(levels(x = group.use2), na.group)
+        group.use2[placeholder.cells] <- na.group
+        cols <- c(hue_pal()(length(x = levels(x = group.use))), "#FFFFFF")
+      } else {
+        cols <- c(hue_pal()(length(x = levels(x = group.use))))
+      }
       pbuild <- ggplot_build(plot = plot)
-      cols <- hue_pal()(length(x = levels(x = group.use)))
-      names(x = cols) <- levels(x = group.use)
-      y.pos <- max(pbuild$layout$panel_params[[1]]$y.range) + 0.25
-      y.max <- y.pos + 0.5
+      names(x = cols) <- levels(x = group.use2)
+      # scale the height of the bar
+      y.range <- diff(x = pbuild$layout$panel_params[[1]]$y.range)
+      y.pos <- max(pbuild$layout$panel_params[[1]]$y.range) + y.range * 0.015
+      y.max <- y.pos + group.bar.height * y.range
+
       plot <- plot + annotation_raster(
-        raster = t(x = cols[sort(x = group.use)]),
+        raster = t(x = cols[group.use2]),,
         xmin = -Inf,
         xmax = Inf,
         ymin = y.pos,
@@ -530,7 +574,9 @@ VlnPlot <- function(
 #' @param object Seurat object
 #' @param dims Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions
 #' @param cells Vector of cells to plot (default is all cells)
-#' @param cols Vector of colors, each color corresponds to an identity class. By default, ggplot2 assigns colors
+#' @param cols Vector of colors, each color corresponds to an identity class. This may also be a single character
+#' or numeric value corresponding to a palette as specified by \code{\link[RColorBrewer]{brewer.pal.info}}.
+#' By default, ggplot2 assigns colors
 #' @param pt.size Adjust point size for plotting
 #' @param reduction Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca
 #' @param group.by Name of one or more metadata columns to group (color) cells by
@@ -593,7 +639,7 @@ DimPlot <- function(
   label.size = 4,
   repel = FALSE,
   cells.highlight = NULL,
-  cols.highlight = 'red',
+  cols.highlight = '#DE2D26',
   sizes.highlight = 1,
   na.value = 'grey50',
   combine = TRUE,
@@ -603,12 +649,7 @@ DimPlot <- function(
   if (length(x = dims) != 2) {
     stop("'dims' must be a two-length vector")
   }
-  reduction <- reduction %||% {
-    default.reductions <- c('umap', 'tsne', 'pca')
-    object.reductions <- FilterObjects(object = object, classes.keep = 'DimReduc')
-    reduc.use <- min(which(x = default.reductions %in% object.reductions))
-    default.reductions[reduc.use]
-  }
+  reduction <- reduction %||% DefaultDimReduc(object = object)
   cells <- cells %||% colnames(x = object)
   data <- Embeddings(object = object[[reduction]])[cells, dims]
   data <- as.data.frame(x = data)
@@ -657,10 +698,10 @@ DimPlot <- function(
         plot <- plot + FacetTheme() +
           facet_wrap(
             facets = vars(!!sym(x = split.by)),
-            ncol = if (length(x = group.by) > 1) {
+            ncol = if (length(x = group.by) > 1 || is.null(x = ncol)) {
               length(x = unique(x = data[, split.by]))
             } else {
-              NULL
+              ncol
             }
           )
       }
@@ -703,15 +744,18 @@ DimPlot <- function(
 #'  may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')
 #' @param split.by A factor in object metadata to split the feature plot by, pass 'ident'
 #'  to split by cell identity'; similar to the old \code{FeatureHeatmap}
+#' @param slot Which slot to pull expression data from?
 #' @param blend Scale and blend expression values to visualize coexpression of two features
 #' @param blend.threshold The color cutoff from weak signal to strong signal; ranges from 0 to 1.
 #' @param ncol Number of columns to combine multiple feature plots to, ignored if \code{split.by} is not \code{NULL}
 #' @param combine Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features
 #' @param coord.fixed Plot cartesian coordinates with fixed aspect ratio
-#' @param by.col If splitting by a factor, plot the splits per column with the features as rows.
+#' @param by.col If splitting by a factor, plot the splits per column with the features as rows; ignored if \code{blend = TRUE}
 
 #'
-#' @return A ggplot object
+#' @return Returns a ggplot object if only 1 feature is plotted.
+#' If >1 features are plotted and \code{combine=TRUE}, returns a combined ggplot object using \code{cowplot::plot_grid}.
+#' If >1 features are plotted and \code{combine=FALSE}, returns a list of ggplot objects.
 #'
 #' @importFrom grDevices rgb
 #' @importFrom cowplot theme_cowplot
@@ -737,7 +781,7 @@ FeaturePlot <- function(
   features,
   dims = c(1, 2),
   cells = NULL,
-  cols = c("lightgrey",  "blue"),
+  cols =  ifelse(test = c(blend, blend), yes = c("#ff0000", "#00ff00"), no = c('lightgrey', 'blue')),
   pt.size = NULL,
   order = FALSE,
   min.cutoff = NA,
@@ -745,10 +789,12 @@ FeaturePlot <- function(
   reduction = NULL,
   split.by = NULL,
   shape.by = NULL,
+  slot = 'data',
   blend = FALSE,
   blend.threshold = 0.5,
   label = FALSE,
   label.size = 4,
+  repel = FALSE,
   ncol = NULL,
   combine = TRUE,
   coord.fixed = FALSE,
@@ -775,10 +821,45 @@ FeaturePlot <- function(
   if (blend && length(x = features) != 2) {
     stop("Blending feature plots only works with two features")
   }
+  if (blend) {
+    if (length(x = cols) > 2) {
+      warning(
+        "Blending feature plots only works with two colors; using first two colors",
+        call. = FALSE,
+        immediate. = TRUE
+      )
+    } else if (length(x = cols) < 2) {
+      warning(
+        "Blended feature plots require two colors, using default colors",
+        call. = FALSE,
+        immediate. = TRUE
+      )
+      cols <- c("#ff0000", "#00ff00")
+    }
+  }
+  if (blend && length(x = cols) != 2) {
+    stop("Blending feature plots only works with two colors")
+  }
   dims <- paste0(Key(object = object[[reduction]]), dims)
   cells <- cells %||% colnames(x = object)
-  data <- FetchData(object = object, vars = c(dims, features), cells = cells)
-  features <- colnames(x = data)[3:ncol(x = data)]
+  data <- FetchData(
+    object = object,
+    vars = c(dims, 'ident', features),
+    cells = cells,
+    slot = slot
+  )
+  if (ncol(x = data) < 4) {
+    stop(
+      "None of the requested features were found: ",
+      paste(features, collapse = ', '),
+      " in slot ",
+      slot,
+      call. = FALSE
+    )
+  } else if (!all(dims %in% colnames(x = data))) {
+    stop("The dimensions requested were not found", call. = FALSE)
+  }
+  features <- colnames(x = data)[4:ncol(x = data)]
   min.cutoff <- mapply(
     FUN = function(cutoff, feature) {
       return(ifelse(
@@ -814,12 +895,12 @@ FeaturePlot <- function(
     yes = brewer.pal.info[cols, ]$maxcolors,
     no = length(x = cols)
   )
-  data[, 3:ncol(x = data)] <- sapply(
-    X = 3:ncol(x = data),
+  data[, 4:ncol(x = data)] <- sapply(
+    X = 4:ncol(x = data),
     FUN = function(index) {
       data.feature <- as.vector(x = data[, index])
-      min.use <- SetQuantile(cutoff = min.cutoff[index - 2], data.feature)
-      max.use <- SetQuantile(cutoff = max.cutoff[index - 2], data.feature)
+      min.use <- SetQuantile(cutoff = min.cutoff[index - 3], data.feature)
+      max.use <- SetQuantile(cutoff = max.cutoff[index - 3], data.feature)
       data.feature[data.feature < min.use] <- min.use
       data.feature[data.feature > max.use] <- max.use
       if (brewer.gran == 2) {
@@ -837,7 +918,7 @@ FeaturePlot <- function(
       return(data.cut)
     }
   )
-  colnames(x = data)[3:ncol(x = data)] <- features
+  colnames(x = data)[4:ncol(x = data)] <- features
   rownames(x = data) <- cells
   data$split <- if (is.null(x = split.by)) {
     RandomName()
@@ -866,7 +947,10 @@ FeaturePlot <- function(
   ylims <- c(floor(min(data[, dims[2]])), ceiling(x = max(data[, dims[2]])))
   if (blend) {
     ncol <- 4
-    color.matrix <- BlendMatrix(col.threshold = blend.threshold)
+    color.matrix <- BlendMatrix(
+      two.colors = cols,
+      col.threshold = blend.threshold
+    )
     colors <- list(
       color.matrix[, 1],
       color.matrix[1, ],
@@ -877,8 +961,16 @@ FeaturePlot <- function(
     ident <- levels(x = data$split)[i]
     data.plot <- data[as.character(x = data$split) == ident, , drop = FALSE]
     if (blend) {
-      data.plot <- cbind(data.plot[, dims], BlendExpression(data = data.plot[, features[1:2]]))
-      features <- colnames(x = data.plot)[3:ncol(x = data.plot)]
+      no.expression <- features[colMeans(x = data.plot[, features]) == 0]
+      if (length(x = no.expression) != 0) {
+        stop(
+          "The following features have no value: ",
+          paste(no.expression, collapse = ', '),
+          call. = FALSE
+        )
+      }
+      data.plot <- cbind(data.plot[, c(dims, 'ident')], BlendExpression(data = data.plot[, features[1:2]]))
+      features <- colnames(x = data.plot)[4:ncol(x = data.plot)]
     }
     for (j in 1:length(x = features)) {
       feature <- features[j]
@@ -889,19 +981,26 @@ FeaturePlot <- function(
         cols.use <- NULL
       }
       plot <- SingleDimPlot(
-        data = data.plot[, c(dims, feature, shape.by)],
+        data = data.plot[, c(dims, 'ident', feature, shape.by)],
         dims = dims,
         col.by = feature,
         order = order,
         pt.size = pt.size,
         cols = cols.use,
         shape.by = shape.by,
-        label = label,
-        label.size = label.size
+        label = FALSE
       ) +
         scale_x_continuous(limits = xlims) +
         scale_y_continuous(limits = ylims) +
         theme_cowplot()
+      if (label) {
+        plot <- LabelClusters(
+          plot = plot,
+          id = 'ident',
+          repel = repel,
+          size = label.size
+        )
+      }
       if (length(x = levels(x = data$split)) > 1) {
         plot <- plot + theme(panel.border = element_rect(fill = NA, colour = 'black'))
         plot <- plot + if (i == 1) {
@@ -937,12 +1036,21 @@ FeaturePlot <- function(
       }
       if (!blend) {
         plot <- plot + guides(color = NULL)
+        cols.grad <- cols
         if (length(x = cols) == 1) {
           plot <- plot + scale_color_brewer(palette = cols)
         } else if (length(x = cols) > 1) {
+          if (all(data.plot[, feature] == data.plot[, feature][1])) {
+            warning("All cells have the same value (", data.plot[1, feature], ") of ", feature, ".")
+            if (data.plot[1, feature][1] == 0) {
+              cols.grad <- cols[1]
+            } else{
+            cols.grad <- cols
+            }
+          }
           plot <- suppressMessages(
             expr = plot + scale_color_gradientn(
-              colors = cols,
+              colors = cols.grad,
               guide = "colorbar"
             )
           )
@@ -1009,13 +1117,19 @@ FeaturePlot <- function(
     } else {
       split.by %iff% 'none'
     }
-    if (by.col & !is.null(x = split.by)) {
-      plots <- lapply(X = plots, FUN = function(x) {
-        suppressMessages(x +
-        theme_cowplot() + ggtitle("") +
-        scale_y_continuous(sec.axis = dup_axis(name = "")) + no.right
-        )
-      })
+    if (by.col && !is.null(x = split.by) && !blend) {
+      plots <- lapply(
+        X = plots,
+        FUN = function(x) {
+          return(suppressMessages(
+            expr = x +
+              theme_cowplot() +
+              ggtitle("") +
+              scale_y_continuous(sec.axis = dup_axis(name = "")) +
+              no.right
+          ))
+        }
+      )
       nsplits <- length(x = levels(x = data$split))
       idx <- 1
       for (i in (length(x = features) * (nsplits - 1) + 1):(length(x = features) * nsplits)) {
@@ -1043,8 +1157,7 @@ FeaturePlot <- function(
         ncol = nsplits,
         legend = legend
       )
-    }
-    else {
+    } else {
       plots <- CombinePlots(
         plots = plots,
         ncol = ncol,
@@ -1185,6 +1298,7 @@ FeatureScatter <- function(
 #' View variable features
 #'
 #' @inheritParams FeatureScatter
+#' @inheritParams HVFInfo
 #' @param cols Colors to specify non-variable/variable status
 #' @param assay Assay to pull variable features from
 #' @param log Plot the x-axis in log scale
@@ -1206,29 +1320,33 @@ VariableFeaturePlot <- function(
   cols = c('black', 'red'),
   pt.size = 1,
   log = NULL,
+  selection.method = NULL,
   assay = NULL
 ) {
   if (length(x = cols) != 2) {
     stop("'cols' must be of length 2")
   }
-  hvf.info <- HVFInfo(object = object, assay = assay)
-  dispersion.names <- c('variance.standardized', 'dispersion.scaled', 'residual_variance')
-  vars <- switch(which(dispersion.names %in% colnames(x = hvf.info)),
-                 c('mean', 'variance.standardized'),
-                 c('mean', 'dispersion'),
-                 c('gmean', 'residual_variance'))
-  axis.labels <- switch(which(dispersion.names %in% colnames(x = hvf.info)),
-                        c('Average Expression', 'Standardized Variance'),
-                        c('Average Expression', 'Dispersion'),
-                        c('Geometric Mean of Expression', 'Residual Variance'))
-  hvf.info <- hvf.info[, vars]
-  log <- log %||% (any(c('variance.standardized', 'residual_variance') %in% colnames(x = hvf.info)))
-  var.features <- VariableFeatures(object = object, assay = assay)
-  var.status <- ifelse(
-    test = rownames(x = hvf.info) %in% var.features,
-    yes = 'yes',
-    no = 'no'
+  hvf.info <- HVFInfo(
+    object = object,
+    assay = assay,
+    selection.method = selection.method,
+    status = TRUE
+    )
+  var.status <- c('no', 'yes')[unlist(x = hvf.info[, ncol(x = hvf.info)]) + 1]
+  hvf.info <- hvf.info[, c(1, 3)]
+  axis.labels <- switch(
+    EXPR = colnames(x = hvf.info)[2],
+    'variance.standardized' = c('Average Expression', 'Standardized Variance'),
+    'dispersion.scaled' = c('Average Expression', 'Dispersion'),
+    'residual_variance' = c('Geometric Mean of Expression', 'Residual Variance')
   )
+  log <- log %||% (any(c('variance.standardized', 'residual_variance') %in% colnames(x = hvf.info)))
+  # var.features <- VariableFeatures(object = object, assay = assay)
+  # var.status <- ifelse(
+  #   test = rownames(x = hvf.info) %in% var.features,
+  #   yes = 'yes',
+  #   no = 'no'
+  # )
   plot <- SingleCorPlot(
     data = hvf.info,
     col.by = var.status,
@@ -1550,10 +1668,11 @@ BarcodeInflectionsPlot <- function(object) {
 #'
 #' Intuitive way of visualizing how feature expression changes across different
 #' identity classes (clusters). The size of the dot encodes the percentage of
-#' cells within a class, while the color encodes the AverageExpression level of
-#' cells within a class (blue is high).
+#' cells within a class, while the color encodes the AverageExpression level
+#' across all cells within a class (blue is high).
 #'
 #' @param object Seurat object
+#' @param assay Name of assay to use, defaults to the active assay
 #' @param features Input vector of features
 #' @param cols Colors to plot, can pass a single character giving the name of
 #' a palette from \code{RColorBrewer::brewer.pal.info}
@@ -1592,6 +1711,7 @@ BarcodeInflectionsPlot <- function(object) {
 #'
 DotPlot <- function(
   object,
+  assay = NULL,
   features,
   cols = c("lightgrey", "blue"),
   col.min = -2.5,
@@ -1605,6 +1725,8 @@ DotPlot <- function(
   scale.max = NA,
   ...
 ) {
+  assay <- assay %||% DefaultAssay(object = object)
+  DefaultAssay(object = object) <- assay
   scale.func <- switch(
     EXPR = scale.by,
     'size' = scale_size,
@@ -1616,6 +1738,9 @@ DotPlot <- function(
     Idents(object = object)
   } else {
     object[[group.by, drop = TRUE]]
+  }
+  if (!is.factor(x = data.features$id)) {
+    data.features$id <- factor(x = data.features$id)
   }
   id.levels <- levels(x = data.features$id)
   data.features$id <- as.vector(x = data.features$id)
@@ -1977,7 +2102,7 @@ VizDimLoadings <- function(
 #' @return A ggplot object
 #'
 #' @importFrom png readPNG
-#' @importFrom ggplot2 ggplot_build ggsave ggplot aes_string geom_blank annotation_raster
+#' @importFrom ggplot2 ggplot_build ggsave ggplot aes_string geom_blank annotation_raster ggtitle
 #'
 #' @export
 #'
@@ -1997,10 +2122,11 @@ AugmentPlot <- function(plot, width = 10, height = 10, dpi = 100) {
     plot = plot,
     geom = class(x = plot$layers[[1]]$geom)[1]
   )
+  title <- plot$labels$title
   tmpfile <- tempfile(fileext = '.png')
   ggsave(
     filename = tmpfile,
-    plot = plot + NoLegend() + NoAxes(),
+    plot = plot + NoLegend() + NoAxes() + theme(plot.title = element_blank()),
     width = width,
     height = height,
     dpi = dpi
@@ -2011,7 +2137,7 @@ AugmentPlot <- function(plot, width = 10, height = 10, dpi = 100) {
     data = plot$data,
     mapping = aes_string(x = xyparams$x, y = xyparams$y)
   ) + geom_blank()
-  blank <- blank + plot$theme
+  blank <- blank + plot$theme + ggtitle(label = title)
   blank <- blank + annotation_raster(
     raster = img,
     xmin = range.values[1],
@@ -2977,28 +3103,50 @@ BlendMap <- function(color.matrix) {
 #
 # @param n Dimensions of blended matrix (n x n)
 # @param col.threshold The color cutoff from weak signal to strong signal; ranges from 0 to 1.
+# @param two.colors Two colors used for the blend expression.
 #
 # @return An n x n matrix of blended colors
 #
-#' @importFrom grDevices rgb
+#' @importFrom grDevices rgb colorRamp
 #
-BlendMatrix <- function(n = 10, col.threshold = 0.5) {
+BlendMatrix <- function(
+  n = 10,
+  col.threshold = 0.5,
+  two.colors=c("#ff0000", "#00ff00")
+) {
   if (0 > col.threshold || col.threshold > 1) {
     stop("col.threshold must be between 0 and 1")
   }
-  return(outer(
-    X = 1:n,
-    Y = 1:n,
-    FUN = function(i, j) {
-      red <- 1 / (1 + exp(x = -(i - col.threshold * n) / 0.9))
-      green <- 1 / (1 + exp(x = -(j - col.threshold * n) / 0.9))
-      blue <- 0.2
-      alpha <- lapply(X = list(red, green, blue), FUN = '^', 40)
-      alpha <- Reduce(f = '+', x = alpha)
-      alpha <- 0.99 - 0.1 * exp(x = -alpha / 1)
-      return(rgb(red = red, green = green, blue = blue, alpha = alpha))
+  ramp <- colorRamp(colors = two.colors)
+  C1 <- ramp(x = 0)
+  C2 <- ramp(x = 1)
+  merge.weight <- min(255 / (C1 + C2 + 0.01))
+  weight_color <- function(w1, c1) {
+    c1_weight <- 1 / (1 + exp(x = -(w1 ^ 1.2 - 3 - col.threshold * 10)))
+    return(c1_weight * c1)
+  }
+  blend_color <- function(i, j) {
+    C1_weight <- weight_color(w1 = i, c1 = C1)
+    C2_weight <- weight_color(w1 = j, c1 = C2)
+    C_blend <- (merge.weight * C1_weight + merge.weight * C2_weight)
+    alpha <- lapply(X = list(i, j), FUN = '^', 0.5)
+    alpha <- Reduce(f = '+', x = alpha)
+    alpha <- (1 - 0.4 /alpha ) * 255
+    return(rgb(
+      red = C_blend[, 1],
+      green = C_blend[, 2],
+      blue = C_blend[, 3],
+      alpha = alpha,
+      maxColorValue = 255
+    ))
+  }
+  blend_matrix <- matrix(nrow = n, ncol = n)
+  for (i in 1:n) {
+    for (j in 1:n) {
+      blend_matrix[i, j] <- blend_color(i = i, j = j)
     }
-  ))
+  }
+  return(blend_matrix)
 }
 
 # Convert R colors to hexadecimal
@@ -3033,6 +3181,59 @@ Col2Hex <- function(...) {
   return(colors)
 }
 
+# Find the default DimReduc
+#
+# Searches for DimReducs matching 'umap', 'tsne', or 'pca', case-insensitive, and
+# in that order. Priority given to DimReducs matching the DefaultAssay or assay specified
+# (eg. 'pca' for the default assay weights higher than 'umap' for a non-default assay)
+#
+# @param object A Seurat object
+# @param assay Name of assay to use; defaults to the default assay of the object
+#
+# @return The default DimReduc, if possible
+#
+DefaultDimReduc <- function(object, assay = NULL) {
+  assay <- assay %||% DefaultAssay(object = object)
+  drs.use <- c('umap', 'tsne', 'pca')
+  dim.reducs <- FilterObjects(object = object, classes.keep = 'DimReduc')
+  drs.assay <- Filter(
+    f = function(x) {
+      return(DefaultAssay(object = object[[x]]) == assay)
+    },
+    x = dim.reducs
+  )
+  if (length(x = drs.assay) > 0) {
+    index <- lapply(
+      X = drs.use,
+      FUN = grep,
+      x = drs.assay,
+      ignore.case = TRUE
+    )
+    index <- Filter(f = length, x = index)
+    if (length(x = index) > 0) {
+      return(drs.assay[min(index[[1]])])
+    }
+  }
+  index <- lapply(
+    X = drs.use,
+    FUN = grep,
+    x = dim.reducs,
+    ignore.case = TRUE
+  )
+  index <- Filter(f = length, x = index)
+  if (length(x = index) < 1) {
+    stop(
+      "Unable to find a DimReduc matching one of '",
+      paste(drs.use[1:(length(x = drs.use) - 1)], collapse = "', '"),
+      "', or '",
+      drs.use[length(x = drs.use)],
+      "', please specify a dimensional reduction to use",
+      call. = FALSE
+    )
+  }
+  return(dim.reducs[min(index[[1]])])
+}
+
 # Plot feature expression by identity
 #
 # Basically combines the codebase for VlnPlot and RidgePlot
@@ -3057,6 +3258,7 @@ Col2Hex <- function(...) {
 # @param ... Extra parameters passed to \code{\link{CombinePlots}}
 #
 #' @importFrom scales hue_pal
+#' @importFrom ggplot2 xlab ylab
 #
 ExIPlot <- function(
   object,
@@ -3110,13 +3312,17 @@ ExIPlot <- function(
     }
     if (is.null(x = cols)) {
       cols <- hue_pal()(length(x = levels(x = idents)))
-    } else if (cols == 'interaction') {
+      cols <- Interleave(cols, InvertHex(hexadecimal = cols))
+    } else if (length(x = cols) == 1 && cols == 'interaction') {
       split <- interaction(idents, split)
       cols <- hue_pal()(length(x = levels(x = idents)))
     } else {
       cols <- Col2Hex(cols)
     }
-    cols <- Interleave(cols, InvertHex(hexadecimal = cols))
+    if (length(x = cols) < length(x = levels(x = split))) {
+      cols <- Interleave(cols, InvertHex(hexadecimal = cols))
+    }
+    cols <- rep_len(x = cols, length.out = length(x = levels(x = split)))
     names(x = cols) <- sort(x = levels(x = split))
   }
   if (same.y.lims && is.null(x = y.max)) {
@@ -3139,6 +3345,28 @@ ExIPlot <- function(
       ))
     }
   )
+  label.fxn <- switch(
+    EXPR = type,
+    'violin' = ylab,
+    'ridge' = xlab,
+    stop("Unknown ExIPlot type ", type, call. = FALSE)
+  )
+  for (i in 1:length(x = plots)) {
+    key <- paste0(unlist(x = strsplit(x = features[i], split = '_'))[1], '_')
+    obj <- names(x = which(x = Key(object = object) == key))
+    if (length(x = obj) == 1) {
+      if (inherits(x = object[[obj]], what = 'DimReduc')) {
+        plots[[i]] <- plots[[i]] + label.fxn(label = 'Embeddings Value')
+      } else if (inherits(x = object[[obj]], what = 'Assay')) {
+        next
+      } else {
+        warning("Unknown object type ", class(x = object), immediate. = TRUE, call. = FALSE)
+        plots[[i]] <- plots[[i]] + label.fxn(label = NULL)
+      }
+    } else if (!features[i] %in% rownames(x = object)) {
+      plots[[i]] <- plots[[i]] + label.fxn(label = NULL)
+    }
+  }
   if (combine) {
     combine.args <- list(
       'plots' = plots,
@@ -3350,13 +3578,13 @@ geom_split_violin <- function(
 #
 InvertHex <- function(hexadecimal) {
   return(vapply(
-    X = hexadecimal,
+    X = toupper(x = hexadecimal),
     FUN = function(hex) {
       hex <- unlist(x = strsplit(
         x = gsub(pattern = '#', replacement = '', x = hex),
         split = ''
       ))
-      key <- unlist(x = strsplit(x = 'FEDCBA9876543210', split = ''))
+      key <- toupper(x = as.hexmode(x = 15:0))
       if (!all(hex %in% key)) {
         stop('All hexadecimal colors must be valid hexidecimal numbers from 0-9 and A-F')
       }
@@ -3495,8 +3723,8 @@ PointLocator <- function(plot, recolor = TRUE, dark.theme = FALSE, ...) {
   points.located <- points.all[which(x = points.all$pip == 1), ]
   #   If we're recoloring, do the recolor
   if (recolor) {
-    no <- ifelse(test = dark.theme, yes = 'white', no = 'black')
-    points.all$color <- ifelse(test = points.all$pip == 1, yes = 'red', no = no)
+    no <- ifelse(test = dark.theme, yes = 'white', no = '#C3C3C3')
+    points.all$color <- ifelse(test = points.all$pip == 1, yes = '#DE2D26', no = no)
     plot.data$color <- points.all$color
     PlotBuild(data = plot.data, dark.theme = dark.theme, ...)
   }
@@ -3782,8 +4010,9 @@ SingleCorPlot <- function(
 # @param dims A two-length numeric vector with dimensions to use
 # @param pt.size Adjust point size for plotting
 # @param col.by ...
-# @param cols Vector of colors, each color corresponds to an identity class. By default, ggplot
-# assigns colors.
+# @param cols Vector of colors, each color corresponds to an identity class. This may also be a single character
+# or numeric value corresponding to a palette as specified by \code{\link[RColorBrewer]{brewer.pal.info}}.
+# By default, ggplot2 assigns colors
 # @param shape.by If NULL, all points are circles (default). You can specify any cell attribute
 # (that can be pulled with FetchData) allowing for both different colors and different shapes on
 # cells.
@@ -3805,6 +4034,7 @@ SingleCorPlot <- function(
 # @param na.value Color value for NA points when using custom scale.
 #
 #' @importFrom cowplot theme_cowplot
+#' @importFrom RColorBrewer brewer.pal.info
 #' @importFrom ggplot2 ggplot aes_string labs geom_text guides
 #' scale_color_brewer scale_color_manual element_rect guide_legend
 #'
@@ -3820,7 +4050,7 @@ SingleDimPlot <- function(
   repel = FALSE,
   label.size = 4,
   cells.highlight = NULL,
-  cols.highlight = 'red',
+  cols.highlight = '#DE2D26',
   sizes.highlight = 1,
   na.value = 'grey50'
 ) {
@@ -3842,7 +4072,7 @@ SingleDimPlot <- function(
       cells.all = rownames(x = data),
       sizes.highlight = sizes.highlight %||% pt.size,
       cols.highlight = cols.highlight,
-      col.base = cols[1] %||% 'black',
+      col.base = cols[1] %||% '#C3C3C3',
       pt.size = pt.size
     )
     order <- highlight.info$plot.order
@@ -3852,7 +4082,7 @@ SingleDimPlot <- function(
     cols <- highlight.info$color
   }
   if (!is.null(x = order) && !is.null(x = col.by)) {
-    if (typeof(x = order) == "logical"){
+    if (typeof(x = order) == "logical") {
       if (order) {
         data <- data[order(data[, col.by]), ]
       }
@@ -3908,7 +4138,7 @@ SingleDimPlot <- function(
     )
   }
   if (!is.null(x = cols)) {
-    plot <- plot + if (length(x = cols) == 1) {
+    plot <- plot + if (length(x = cols) == 1 && (is.numeric(x = cols) || cols %in% rownames(x = brewer.pal.info))) {
       scale_color_brewer(palette = cols, na.value = na.value)
     } else {
       scale_color_manual(values = cols, na.value = na.value)
@@ -4175,7 +4405,7 @@ SingleRasterMap <- function(
   plot <- ggplot(data = data) +
     my_geom(mapping = aes_string(x = 'Cell', y = 'Feature', fill = 'Expression')) +
     theme(axis.text.x = element_blank(), axis.ticks.x = element_blank()) +
-    scale_fill_gradientn(limits = limits, colors = colors) +
+    scale_fill_gradientn(limits = limits, colors = colors, na.value = "white") +
     labs(x = NULL, y = NULL, fill = group.by %iff% 'Expression') +
     WhiteBackground() + NoAxes(keep.text = TRUE)
   if (!is.null(x = group.by)) {