+++ NEWS.md
@@ -2,6 +2,44 @@
 All notable changes to Seurat will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [3.0.2] - 2019-06-07
+### Added
+- Flag to skip singleton grouping in `FindClusters`
+- New custom colors for blended `FeaturePlot`s
+- New `GetResidual` function
+- New Seurat/Monocle converters
+
+### Changed
+- Fix issue where certain assays weren't being shown in the `Seurat` object
+- Fix issue where we weren't updating `DimReduc` object column names
+- Fix line spacers in `DoHeatmap`
+- Fix uninformative labels in `FeaturePlot`
+- Fix unset identities when converting from SCE to Seurat
+- Fix single colors being interpreted as palettes in `SingleDimPlot`
+- Ensure factor levels are always numerically increasing after `FindClusters`
+- Better cell highlighting colors for `DimPlot`
+- Fix to `levels<-.Seurat`
+- Add ability to use counts/scaled data in `BuildClusterTree`
+- Minor fix to split `ScaleData`
+
+## [3.0.1] - 2019-05-16
+### Added
+- Add global option (Seurat.memsafe) to skip gc() calls
+- Restore draw.lines to DoHeatmap, maintain size of color bar with different number of features (#1429)
+- Enable split.by parameter for ScaleData
+- Add slot parameter to FeaturePlot (#1483)
+- Add assay parameter to DotPlot (#1404)
+
+### Changed
+- Fix to color options for VlnPlot with split.by option (#1425)
+- Improvements to conversion functions (loom, SCE)
+- Fix for cluster tree reordering (#1434)
+- Fix PercentageFeatureSet for single feature case
+- Fix to fold change calculation and filtering for other slots in FindMarkers (#1454)
+- Keep title vectorized in AugmentPlot (#1515)
+- Export LogSeuratCommand function
+- Fix for FindConservedMarkers when one ident is missing from a group (#1517)
+
 ## [3.0.0] - 2019-04-16
 ### Added
 - New method for identifying anchors across single-cell datasets